Proof-of-concept legacy code abstraction
