+++ text-serializer-legacy/src/main/java/net/kyori/adventure/text/serializer/legacy/FormatCodeType.java
@@ -1,0 +1,7 @@
+package net.kyori.adventure.text.serializer.legacy;
+
+enum FormatCodeType {
+  MOJANG_LEGACY,
+  KYORI_HEX,
+  BUNGEECORD_UNUSUAL_HEX;
+}
